requirements.txt
apng==0.3.4
astroid==2.12.5
autopep8==1.7.0
colorama==0.4.5
dill==0.3.5.1
euclid3==0.1
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
numpy==1.23.2
openscad-runner==1.1.0
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
prettytable==0.7.2
pycodestyle==2.9.1
pygifsicle==1.0.6
pylint==2.15.0
pypng==0.0.19
PyYAML==6.0
qrcode==7.3.1
solidpython==1.1.3
support-developer==1.0.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
